fix(imports): split out templating from ./utils
This should fix https://github.com/algolia/react-instantsearch/issues/3632, as renderTemplate will no longer be in the imported path for connectors (react instantsearch hooks), and thus no errors can be in it either
