+++ src/components/Breadcrumb/Breadcrumb.tsx
@@ -8,7 +8,7 @@ import type {
   BreadcrumbTemplates,
 } from '../../widgets/breadcrumb/breadcrumb';
 import type { ComponentCSSClasses } from '../../types';
-import type { PreparedTemplateProps } from '../../lib/utils';
+import type { PreparedTemplateProps } from '../../lib/templating';
 import type { BreadcrumbConnectorParamsItem } from '../../connectors/breadcrumb/connectBreadcrumb';
 
 export type BreadcrumbComponentCSSClasses =
