+++ src/components/Breadcrumb/__tests__/Breadcrumb-test.tsx
@@ -7,7 +7,7 @@ import { h } from 'preact';
 import { render, fireEvent } from '@testing-library/preact';
 import type { BreadcrumbProps } from '../Breadcrumb';
 import Breadcrumb from '../Breadcrumb';
-import { prepareTemplateProps } from '../../../lib/utils';
+import { prepareTemplateProps } from '../../../lib/templating';
 import defaultTemplates from '../../../widgets/breadcrumb/defaultTemplates';
 
 const defaultProps: BreadcrumbProps = {
