+++ src/components/GeoSearchControls/__tests__/GeoSearchControls-test.tsx
@@ -6,7 +6,7 @@
 import { h } from 'preact';
 import { render, fireEvent } from '@testing-library/preact';
 import GeoSearchControls from '../GeoSearchControls';
-import { prepareTemplateProps } from '../../../lib/utils';
+import { prepareTemplateProps } from '../../../lib/templating';
 
 describe('GeoSearchControls', () => {
   const CSSClassesDefaultProps = {
