+++ src/components/Template/Template.tsx
@@ -2,13 +2,11 @@
 
 import type { JSX } from 'preact';
 import { h, Component } from 'preact';
-import { warning, renderTemplate, isEqual } from '../../lib/utils';
+import { warning, isEqual } from '../../lib/utils';
+import { renderTemplate } from '../../lib/templating';
 
-import type {
-  BindEventForHits,
-  SendEventForHits,
-  PreparedTemplateProps,
-} from '../../lib/utils';
+import type { BindEventForHits, SendEventForHits } from '../../lib/utils';
+import type { PreparedTemplateProps } from '../../lib/templating';
 import type { Templates } from '../../types';
 
 const defaultProps = {
