+++ src/helpers/components/ReverseHighlight.tsx
@@ -2,12 +2,13 @@
 import { h } from 'preact';
 
 import { ReverseHighlight as ReverseHighlightUiComponent } from '../../components/ReverseHighlight/ReverseHighlight';
-// These utils are individually imported, as utils/renderTemplate imports helpers/components, importing lib/utils would create a circular dependency.
-import { getPropertyByPath } from '../../lib/utils/getPropertyByPath';
-import { unescape } from '../../lib/utils/escape-html';
-import { toArray } from '../../lib/utils/toArray';
-import { warning } from '../../lib/utils/logger';
-import { getHighlightedParts } from '../../lib/utils/getHighlightedParts';
+import {
+  getPropertyByPath,
+  unescape,
+  toArray,
+  warning,
+  getHighlightedParts,
+} from '../../lib/utils';
 
 import type {
   BaseHit,
