+++ src/helpers/components/Snippet.tsx
@@ -2,12 +2,13 @@
 import { h } from 'preact';
 
 import { Snippet as SnippetUiComponent } from '../../components/Snippet/Snippet';
-// These utils are individually imported, as utils/renderTemplate imports helpers/components, importing lib/utils would create a circular dependency.
-import { getPropertyByPath } from '../../lib/utils/getPropertyByPath';
-import { toArray } from '../../lib/utils/toArray';
-import { unescape } from '../../lib/utils/escape-html';
-import { warning } from '../../lib/utils/logger';
-import { getHighlightedParts } from '../../lib/utils/getHighlightedParts';
+import {
+  getPropertyByPath,
+  unescape,
+  toArray,
+  warning,
+  getHighlightedParts,
+} from '../../lib/utils';
 
 import type {
   BaseHit,
