+++ src/lib/templating/__tests__/prepareTemplateProps.test.ts
