+++ src/lib/templating/__tests__/renderTemplate.test.ts
