+++ src/lib/templating/index.ts
@@ -1,0 +1,2 @@
+export * from './prepareTemplateProps';
+export * from './renderTemplate';
