+++ src/lib/templating/prepareTemplateProps.ts
@@ -1,4 +1,4 @@
-import { uniq } from './uniq';
+import { uniq } from '../utils/uniq';
 import type { HoganHelpers, Templates } from '../../types';
 import type { HoganOptions } from 'hogan.js';
