+++ src/lib/utils/index.ts
@@ -37,9 +37,7 @@ export * from './isSpecialClick';
 export * from './logger';
 export * from './mergeSearchParameters';
 export * from './noop';
-export * from './prepareTemplateProps';
 export * from './range';
-export * from './renderTemplate';
 export * from './resolveSearchParameters';
 export * from './reverseHighlightedParts';
 export * from './safelyRunOnBrowser';
