+++ src/widgets/geo-search/createHTMLMarker.ts
@@ -1,6 +1,6 @@
 /* global google EventListener */
 import { render } from 'preact';
-import type { renderTemplate } from '../../lib/utils';
+import type { renderTemplate } from '../../lib/templating';
 
 export type HTMLMarkerArguments = {
   __id: string;
