+++ src/widgets/hierarchical-menu/hierarchical-menu.tsx
@@ -11,12 +11,12 @@ import type {
 } from '../../connectors/hierarchical-menu/connectHierarchicalMenu';
 import connectHierarchicalMenu from '../../connectors/hierarchical-menu/connectHierarchicalMenu';
 import defaultTemplates from './defaultTemplates';
-import type { PreparedTemplateProps } from '../../lib/utils';
+import type { PreparedTemplateProps } from '../../lib/templating';
 import {
-  prepareTemplateProps,
   getContainerNode,
   createDocumentationMessageGenerator,
 } from '../../lib/utils';
+import { prepareTemplateProps } from '../../lib/templating';
 import type {
   TransformItems,
   Template,
