+++ src/widgets/refinement-list/refinement-list.tsx
@@ -11,13 +11,13 @@ import type {
 } from '../../connectors/refinement-list/connectRefinementList';
 import connectRefinementList from '../../connectors/refinement-list/connectRefinementList';
 import {
-  prepareTemplateProps,
   getContainerNode,
   createDocumentationMessageGenerator,
 } from '../../lib/utils';
+import { prepareTemplateProps } from '../../lib/templating';
 import { component } from '../../lib/suit';
 import type { Template, WidgetFactory, Renderer } from '../../types';
-import type { PreparedTemplateProps } from '../../lib/utils';
+import type { PreparedTemplateProps } from '../../lib/templating';
 import searchBoxDefaultTemplates from '../search-box/defaultTemplates';
 import type { SearchBoxTemplates } from '../search-box/search-box';
 import type { SearchBoxComponentTemplates } from '../../components/SearchBox/SearchBox';
